Fix: route && user && vehicle_add/_list screens for mobile; route_update screen
