+++ ePrijevozSarajevo/UI/eprijevoz_desktop/lib/layouts/master_screen.dart
@@ -4,11 +4,10 @@ import 'package:eprijevoz_desktop/providers/auth_provider.dart';
 import 'package:eprijevoz_desktop/providers/user_provider.dart';
 import 'package:eprijevoz_desktop/screens/home_screen.dart';
 import 'package:eprijevoz_desktop/screens/request/request_list_screen.dart';
-import 'package:eprijevoz_desktop/screens/route_list_screen.dart';
+import 'package:eprijevoz_desktop/screens/route/route_list_screen.dart';
 import 'package:eprijevoz_desktop/screens/statistic_screen.dart';
 import 'package:eprijevoz_desktop/screens/user/user_list_screen.dart';
-//import 'package:eprijevoz_desktop/screens/user_list_screen.dart';
-import 'package:eprijevoz_desktop/screens/vehicle_list_screen.dart';
+import 'package:eprijevoz_desktop/screens/vehicle/vehicle_list_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:provider/provider.dart';
